requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
DataProperty==0.50.1
docker==4.4.4
Flask==1.1.2
humanreadable==0.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kaggle==1.5.10
kiwisolver==1.3.1
loguru==0.5.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mbstrdecoder==1.0.1
msgfy==0.1.0
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.2
path==15.1.2
path.py==12.5.0
pathvalidate==2.4.1
Pillow==8.1.0
pthflops==0.3.5
pyparsing==2.4.7
pyroute2==0.5.17
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
pywin32==227
requests==2.25.1
scipy==1.6.0
SimpleSQLite==1.1.4
six==1.15.0
sqliteschema==1.0.5
subprocrunner==1.2.1
tabledata==1.1.3
tcconfig==0.26.0
text-unidecode==1.3
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.59.0
typepy==1.1.4
typing-extensions==3.7.4.3
urllib3==1.26.3
voluptuous==0.12.1
websocket-client==0.58.0
Werkzeug==1.0.1
win-inet-pton==1.1.0
win32-setctime==1.0.3
